WIP: major fixes in a variety of functions

+++ .bash_profile
@@ -1,52 +1,41 @@
 umask 022
 
-source "$HOME"/.functions
-
-#----------
-#TODO have a Hash that defines $bin, $bindir and search paths specific to OS
-#     and reduce this to a nested FOR loop (see bashrc.$prog)
-#----------
-
-addPath PATH -"$HOME"/{.local/,}bin
 export PATH MANPATH
 export LANG='en_US.utf8'
 
-if [[ "$-" == *i* ]] || tty -s; then
-  if which ssh-agent &>/dev/null; then
-    if [ -z "$SSH_AUTH_SOCK" ]; then
-        eval `ssh-agent ${SSH_AGENT_ARGS:-${BASH_VERSION:+ -s}}`
-        trap "kill -9 $SSH_AGENT_PID" EXIT
-    fi
-    [ -n "$SSH_AUTH_SOCK" ] && ssh-add -q "$HOME"/.ssh/{id_?sa,*.pem} 2>/dev/null
-
-    # CAC/PIF card support
-    case ${OSTYPE:-`uname`} in
-      [lL]inux*)
-            : ${OPENSC_LIB:=/usr/lib/`uname -m`-linux-gnu/opensc-pkcs11.so} ;;
-      cygwin|CYGWIN_*)
-            # MSI doesn't allow tailoring path
-            : ${OPENSC_LIB:=`convert_path "$PROGRAMFILES/OpenSC Project/OpenSC/pkcs11/opensc-pkcs11.dll"`} ;;
-      [dD]arwin*)
-            : ${OPENSC_LIB:=/usr/local/lib/opensc-pkcs11.so}
-    esac
-    if [ -f "$OPENSC_LIB" ]; then
-        export OPENSC_LIB
-        [ -n "$SSH_AUTH_SOCK" ] && ssh-add -qs "$OPENSC_LIB" 2>/dev/null
-    else
-        unset OPENSC_LIB
-    fi
-  fi
-fi
-
-for f in "$HOME"/.{bash_profile.local,bashrc_${OSTYPE:=`uname`}}; do
-  [ -f "$f" ] || continue
-  source "$f"
-done
-
 : ${EDITOR:=`which "$EDITOR" vim vi nano pico emacs 2>/dev/null | head -n 1`}
-: ${PAGER:='less -RSF'}
+: ${PAGER:='less -RF'}
 export EDITOR PAGER
 
-source "$HOME"/.bashrc
+if [ -z "$SSH_AUTH_SOCK" -a `which ssh-agent 2>/dev/null` ]; then
+  eval `ssh-agent ${SSH_AGENT_ARGS:-${BASH_VERSION:+ -s}}`
+  trap "kill -9 $SSH_AGENT_PID" EXIT
+fi
+if [[ "$-" == *i* ]] || tty -s; then
+  #FIXME technically could pipe in password from file
+  [ -n "$SSH_AUTH_SOCK" ] && ssh-add "$HOME"/.ssh/{id_?sa,*.pem} 2>/dev/null
+fi
+
+echo "IFS before anything. "
+cat -etv <<<"$IFS"
+
+for f in "$HOME"/.{bash_profile.local,bashrc}; do
+  [ -f "$f" ] || continue
+  source "$f" || echo >&2 "RC=$? in $f"
+done
+unset f
+
+# CAC/PIF card support
+#FIXME paths should come from bashrc_`uname`, not hard-coded here
+#case ${OSTYPE:-`uname`} in
+#  [lL]inux*)
+#        : ${OPENSC_LIB:=/usr/lib/`uname -m`-linux-gnu/opensc-pkcs11.so} ;;
+#  cygwin|CYGWIN_*)
+#        # MSI doesn't allow tailoring path
+#        : ${OPENSC_LIB:=`convert_path "$PROGRAMFILES/OpenSC Project/OpenSC/pkcs11/opensc-pkcs11.dll"`} ;;
+##  [dD]arwin*)
+#        : ${OPENSC_LIB:=/usr/local/lib/opensc-pkcs11.so}
+#esac
+#[ -f "$OPENSC_LIB" -a  -n "$SSH_AUTH_SOCK" ] && ssh-add -s "$OPENSC_LIB" 2>/dev/null
 
 # vim: expandtab:ts=4:sw=4
